% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSKmeans.R
\name{SSKmeans}
\alias{SSKmeans}
\title{SSKmeans}
\usage{
SSKmeans(x, K = NULL, lambda1 = 0, lambda2 = 0, E = NULL, nstart = 20,
  silent = FALSE, maxiter = 20)
}
\arguments{
\item{x}{a data matrix of dimension n * p, where n is number of samples to be clustered, p is number of features.}

\item{K}{pre-specified number of clusters}

\item{lambda1}{tuning parameter for l1 norm lasso penalty. Large lambda1 will induce more feature weights to be 0.}

\item{E}{network structure. E should be a m by 2 matrix, where m is total number of connections (edges) in the graph. 
For example, if feature 2 and feature 3 are connected in the graph, E[j,] <- c(2, 3). j = 1, ..., m.}

\item{nstart}{number of initialization for Kmeans.}

\item{silent}{if print out progress.}

\item{maxiter}{max number of iterations.}

\item{lambda1}{tuning parameter for the smoothness of feature selection. Large lambda will induce feature weights to be similar.}
}
\value{
a list. The following items are included in the list.
\item{ws}{weight for each feature. Zero weight means the feature is not selected.}
\item{Cs}{Cluster Assignment}
\item{wcss}{within cluster sum of square}
\item{crit}{objective value}
\item{E}{network structure}
}
\description{
Smoothed sparse Kmeans
}
\details{
Perform sparse Kmeans to perform sample clustering and feature selection. In feature selection, we also want to incorporate spatio information such that adjacent voxels have similar coefficient.
}
\examples{
set.seed(11)
x <- matrix(rnorm(50*70),ncol=70)
x[1:25,1:20] <- x[1:25,1:20]+1
x <- scale(x, TRUE, TRUE)
# choose tuning parameter
lambda1 <- 1
lambda2 <- 1
E <- cbind(1:69,2:70)
K=2
nstart=20
silent=FALSE
maxiter=6

km.out <- SSKmeans(x,K=2,lambda1=1, lambda2=1, E=E)

}
\author{
Caleb
}
